refactor(site): reduce nested components for the workspaces table
