+++ site/src/components/TableEmpty/TableEmpty.tsx
@@ -1,4 +1,3 @@
-import { makeStyles } from "@mui/styles"
 import TableCell from "@mui/material/TableCell"
 import TableRow from "@mui/material/TableRow"
 import { FC } from "react"
@@ -10,19 +9,11 @@ import {
 export type TableEmptyProps = EmptyStateProps
 
 export const TableEmpty: FC<TableEmptyProps> = (props) => {
-  const styles = useStyles()
-
   return (
     <TableRow>
-      <TableCell colSpan={999} className={styles.tableCell}>
+      <TableCell colSpan={999} sx={{ padding: "0 !important" }}>
         <EmptyState {...props} />
       </TableCell>
     </TableRow>
   )
 }
-
-const useStyles = makeStyles(() => ({
-  tableCell: {
-    padding: "0 !important",
-  },
-}))
