+++ site/src/pages/WorkspacesPage/WorkspacesPageView.tsx
@@ -12,7 +12,7 @@ import {
 } from "components/PageHeader/PageHeader"
 import { Stack } from "components/Stack/Stack"
 import { WorkspaceHelpTooltip } from "components/Tooltips"
-import { WorkspacesTable } from "components/WorkspacesTable/WorkspacesTable"
+import { WorkspacesTable } from "pages/WorkspacesPage/WorkspacesTable"
 import { useLocalStorage } from "hooks"
 import difference from "lodash/difference"
 import { ImpendingDeletionBanner, Count } from "components/WorkspaceDeletion"
